+++ backend/src/main/java/aiversum/backend/rest/dto/VertexaiModelResponse.java
@@ -1,0 +1,8 @@
+package aiversum.backend.rest.dto;
+
+import java.util.List;
+
+public record VertexaiModelResponse(List<VertexaiModel> models) {
+    public record VertexaiModel(String name){
+    }
+}